+++ example/lib/main.dart
@@ -1,6 +1,3 @@
-import 'dart:developer';
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'dart:async';
 import 'package:fcm_channels_manager/fcm_channels_manager.dart';
@@ -26,13 +23,7 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
     WidgetsBinding.instance.addPostFrameCallback((_) {
       initChannels();
     });
-    if (Platform.isIOS) {
-      _fcmChannelsManagerPlugin
-          .requestNotificationPermission(
-            providesAppNotificationSettings: true,
-          )
-          .then((value) => log('Permission status: $value'));
-    }
+
     WidgetsBinding.instance.addObserver(this);
   }