+++ CHANGELOG.md
@@ -19,3 +19,17 @@
 Improvent request notification permission for iOS
 * New `requestNotificationPermission` function support full options for iOS
 * Get current permission status with `getNotificationStatus`
+
+## 0.0.6
+
+Add new function to provide app notification settings
+
+* Add: `providesAppNotificationSettings` return true if this option added
+
+Remove request notification permission for iOS from version 0.0.5
+* Remove: `requestNotificationPermission`
+* Remove: `getNotificationStatus`
+
+## 0.0.7
+
+* Fix bug can't run plugin